+++ packages/tools/utils/CHANGELOG.md
@@ -4,6 +4,7 @@
 - added `isRegexpEmpty`, `isRegexpNotEmpty`
 - added `makeRegexOf`, `makeSafeRegexOf`
 - added `regexOf`, `safeRegexOf`
+- changed: `arrayAverage`, `keyValueArrayAverage`, `makeAverageWith` return zero when passed empty arrays
 - updated license year
 
 ## `@svizzle/utils` v0.18.0